+++ alexflipnote/client.py
@@ -51,13 +51,13 @@ def _parse_text(text: str) -> str:
 
 class Client:
 
-    def __init__(self) -> None:
-        self._session = ClientSession(loop = get_event_loop())
+    def __init__(self, session: ClientSession = None) -> None:
+        self.session = ClientSession(loop = get_event_loop()) or session
         self._api_url = "https://api.alexflipnote.dev"
 
     # will be rewritten when json errors are a thing.
     async def _check_url(self, url: str):  # return_response=False):
-        response = await self._session.get(url)
+        response = await self.session.get(url)
         # print(response.status, await response.text())
         # error.group().strip("</p>")
         if response.status in [400, 404, 500]:
@@ -82,7 +82,7 @@ class Client:
     # Json/URL
 
     async def support_server(self, creator: bool = False) -> Tuple[Any, str]:
-        api = await self._session.get(self._api_url)
+        api = await self.session.get(self._api_url)
         support_server = (await api.json()).get("support_server")
         if creator:
             return support_server, "https://discord.gg/yCzcfju"
@@ -90,24 +90,24 @@ class Client:
         return support_server
 
     async def birb(self) -> str:
-        response = await self._session.get(f"{self._api_url}/birb")
+        response = await self.session.get(f"{self._api_url}/birb")
 
         return (await response.json()).get('file')
 
     async def cats(self) -> str:
-        response = await self._session.get(f"{self._api_url}/cats")
+        response = await self.session.get(f"{self._api_url}/cats")
         return (await response.json()).get('file')
 
     async def sadcat(self) -> str:
-        response = await self._session.get(f"{self._api_url}/sadcat")
+        response = await self.session.get(f"{self._api_url}/sadcat")
         return (await response.json()).get('file')
 
     async def fml(self) -> str:
-        response = await self._session.get(f"{self._api_url}/fml")
+        response = await self.session.get(f"{self._api_url}/fml")
         return (await response.json()).get('text')
 
     async def dogs(self) -> str:
-        response = await self._session.get(f"{self._api_url}/dogs")
+        response = await self.session.get(f"{self._api_url}/dogs")
         return (await response.json()).get('file')
 
     # Colour
@@ -119,7 +119,7 @@ class Client:
         if not search(r'^(?:[0-9a-fA-F]{3}){1,2}$', colour):
             raise BadRequest("Invalid HEX value. You're only allowed to enter HEX (0-9 & A-F)")
 
-        response = await self._session.get(f"{self._api_url}/colour/{colour}")
+        response = await self.session.get(f"{self._api_url}/colour/{colour}")
         color = await response.json()
 
         return Colour(color)
@@ -127,7 +127,7 @@ class Client:
     # Steam
 
     async def steam(self, profile: str) -> Steam:
-        response = await self._session.get(f"{self._api_url}/steam/user/{profile}")
+        response = await self.session.get(f"{self._api_url}/steam/user/{profile}")
         if response.status == 404:  # maybe remove and let _check_url handle it.
             raise NotFound("User not found on steam.")
 
@@ -138,7 +138,7 @@ class Client:
     # Dict
 
     async def github_colours(self) -> dict:
-        response = await self._session.get(f"{self._api_url}/color/github")
+        response = await self.session.get(f"{self._api_url}/color/github")
         colors = await response.json()
 
         return dict(colors)
@@ -169,33 +169,33 @@ class Client:
 
         url = f"{self._api_url}/achievement?text={text}{actual_icon}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def amiajoke(self, image: str) -> Image:
         url = await self._check_url(
             f"{self._api_url}/amiajoke?image={image}"
             )
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def bad(self, image: str) -> Image:
         url = await self._check_url(
             f"{self._api_url}/bad?image={image}"
             )
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def calling(self, text: str) -> Image:
         text = _parse_text(text)
         url = f"{self._api_url}/calling?text={text}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def captcha(self, text: str) -> Image:
         text = _parse_text(text)
         url = f"{self._api_url}/captcha?text={text}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def challenge(self, text: str, icon: Union[int, str, Icon] = None) -> Image:
         text = _parse_text(text)
@@ -221,7 +221,7 @@ class Client:
 
         url = f"{self._api_url}/challenge?text={text}{actual_icon}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def colour_image(self, colour = None) -> Image:
         if colour is None:
@@ -232,7 +232,7 @@ class Client:
 
         url = f"{self._api_url}/colour/image/{colour}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def colour_image_gradient(self, colour = None) -> Image:
         if colour is None:
@@ -243,7 +243,7 @@ class Client:
 
         url = f"{self._api_url}/colour/image/gradient/{colour}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def colourify(self, image: str, colour = "", background = "") -> Image:
         if colour != "":
@@ -260,27 +260,27 @@ class Client:
 
         url = f"{self._api_url}/colourify?image={image}{colour}{background}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def didyoumean(self, top: str, bottom: str) -> Image:
         top = _parse_text(top)
         bottom = _parse_text(bottom)
         url = f"{self._api_url}/didyoumean?top={top}&bottom={bottom}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def drake(self, top: str, bottom: str) -> Image:
         top = _parse_text(top)
         bottom = _parse_text(bottom)
         url = f"{self._api_url}/drake?top={top}&bottom={bottom}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def facts(self, text: str) -> Image:
         text = _parse_text(text)
         url = f"{self._api_url}/facts?text={text}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def filter(self, name: str, image: str) -> Image:
         options = ['blur', 'invert', 'b&w', 'deepfry', 'sepia', 'pixelate',
@@ -293,7 +293,7 @@ class Client:
             name = choice(options)
 
         url = await self._check_url(f"{self._api_url}/filter/{name}?image={image}")
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def floor(self, text: str, image: str = None) -> Image:
         text = _parse_text(text)
@@ -302,41 +302,41 @@ class Client:
 
         url = f"{self._api_url}/floor?text={text}{image}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def jokeoverhead(self, image: str) -> Image:
         url = await self._check_url(
             f"{self._api_url}/jokeoverhead?image={image}"
             )
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def pornhub(self, text: str, text2: str) -> Image:
         text = _parse_text(text)
         text2 = _parse_text(text2)
         url = f"{self._api_url}/pornhub?text={text}&text2={text2}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def salty(self, image: str) -> Image:
         url = await self._check_url(
             f"{self._api_url}/salty?image={image}"
             )
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def scroll(self, text: str) -> Image:
         text = _parse_text(text)
         url = f"{self._api_url}/scroll?text={text}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def ship(self, user: str, user2: str) -> Image:
         url = await self._check_url(
             f"{self._api_url}/ship?user={user}&user2={user2}"
             )
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def supreme(self, text: str, dark: bool = False, light: bool = False) -> Image:
         text = _parse_text(text)
@@ -350,14 +350,14 @@ class Client:
 
         url = f"{self._api_url}/supreme?text={text}{darkorlight}"
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     async def trash(self, face: str, trash: str) -> Image:
         url = await self._check_url(
             f"{self._api_url}/trash?face={face}&trash={trash}"
             )
 
-        return Image(url, self._session)
+        return Image(url, self.session)
 
     # aliases
 
@@ -366,5 +366,5 @@ class Client:
     color_image = colour_image
 
     async def close(self) -> None:
-        if not self._session.closed:
-            await self._session.close()
+        if not self.session.closed:
+            await self.session.close()